# Lab2/src/main/java/com/example/lab2/entities/Student.java
package com.example.lab2.entities;

public class Student {
    private int id;
    private String name;
    private String email;

    public Student() {

    }
    public Student(int id, String name) {
        this.id = id;
        this.name = name;
    }

    public int getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public String getEmail() {
        return email;
    }

    public void setId(int id) {
        this.id = id;
    }

    public void setName(String name) {
        this.name = name;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}